Refactor : Referee클래스에 있던 calculateStrikeAndBall() Hint클래스로 이동
